Split generate_static_data.py into focused modules
Move the 1213-line script into a scripts/static_data/ package with one
module per responsibility: business_days, loaders, filters, enrichment,
and one outputs/*.py per generated JSON artefact. generate_static_data.py
is now a ~130-line orchestrator that loads source data, enriches mergers
once, then calls each generator.

filters.py centralises the waiver/notification split that was duplicated
inline in generate_stats_json and generate_analysis_json, and exposes
exclude_for_public_output() (waivers + suspended) used by upcoming events.

Output is byte-identical to the pre-refactor script, verified against a
snapshot of merger-tracker/frontend/public/data/ and data/output/.

Also adds scripts/tests/test_static_data_filters.py and
test_static_data_outputs.py covering filters and each generator.
